edited createIndexTables to calculate weights correctly - include all weights of bonds between 0 and 100 TTMs, as the weighting of 70 buckets to 6 buckets includes the remaining universe weights (the buckets in bucket 70 that are included in bucket 6, eg bucket 69, bucket 70, etc) plus year 35.25 to year 100) as the weights of the 6th bucket.Where, the cashflows are generated from bonds < 35.25 years as that is our investment.
Since the weights of all weights from 35 to 100 are included in the bucket #6 as well, then for IndexTable we include all weights in over_35 rather than only the weights <35.25 or <35 years, which equals to 1, hence not dividing by it anymore.

we do not need to normalize by only the weights <=35.25 or 35 years.